// WebApp/src/app/utils/enums.ts
export enum ONBOARDING_STEPS {
    GET_STARTED = 0,
    YOUTUBE_CHANNEL = 1,
    REVENUE_ESTIMATOR = 2,
    INTEGRATION = 3,
    TERMS_CONDITIONS = 4,
    WELCOME = 5
}

export enum PROVIDERS_STATUS{
    OPERATIONAL = "OPERATIONAL",
    DELAYED = "DELAYED",
}

export enum REFERRAL_SOURCE {
    FACEBOOK = "facebook",
    INSTAGRAM = "Instagram",
    EMAIL = "Email",
}

export enum REFERRAL_MEDIUM {
    SOCIAL = "social",
    CPC = "cpc",
    ORGANIC = "organic",
}

export enum REFERRAL_CAMPAIGN {
    SUMMER_2025 = "Summer 2025",
    LAUNCH_PROMO = "launch_promo"
}

export enum FAN_MESSAGE_TYPE {
  LIVESTREAM = 'livestream',
  VIDEO = 'video',
}

export enum TRANSACTION_REVENUE_STATUS {
  AwaitingReply = 'awaiting-reply',
  Earned = 'earned',
  AutoReplied = 'auto-replied',
}

export enum REVENUE_STATUS {
  AwaitingReply = 'awaiting-reply',
  Earned = 'earned',
  AutoReplied = 'auto-replied',
}